Make blockMarkers use "bytes of UTF8" instead of "UCS2 code points"
As suggested in PR feedback I have switched the offset mechanism for the
`blockMarkers` property to indicate the number of bytes required to
represent the given `innerHTML` as a UTF8 string.

This shifts computational burden from PHP back to JavaScript. It
simplifies the requirements on the server since it need only use
the built-in `strlen()` and `substr()` family of functions (instead
of the `mb_strlen()` and `iconv()` friends). On the JavaScript side
it requires providing a function which can compute those byte
lenths from internal JavaScript strings.

Right now there is no reassembly of broken-apart documents but in
the HTML of a parse from within JavaScript we'll need to provide a
way to compute a running length number of bytes and then convert that
back into JavaScript string indices; this shouldn't be too obstructive.

After these changes we'll be standardized on byte-length which requires
fewer assumptions than UC2 code points but it does require that we
compare against a UTF8 string, which I think we need to make sure is
clear.